requirements.txt
aiohttp==3.6.2
astroid==2.3.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.5.2
pylint==2.4.3
python-dateutil==2.8.0
requests==2.22.0
schedule==0.6.0
six==1.12.0
slackclient==2.2.1
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.3.0